feat(architecture): Extract AI cache optimization logic to service layer
- Created AICacheOptimizationService in lib/services/
- Refactored route handler to delegate to service layer
- Reduced route file from 335 to 30 lines (91% reduction)
- Removed duplicate PerformanceMetrics interface
- Maintained perfect service layer compliance (blueprint.md:208-209)
- All quality gates passing (security, lint, typecheck, build)
