+++ app/api/performance/ai-cache-optimization/route.ts
@@ -4,12 +4,12 @@
  * Provides real-time metrics and analytics for AI cache optimization,
  * showing cost savings, performance improvements, and cache efficiency.
  *
- * Refactored to use standardized APIRouteHandler pattern
+ * Uses standardized APIRouteHandler pattern with service layer delegation.
  */
 
 import { logger } from "@/lib/logger";
-import { redisManager } from "@/lib/redis";
 import { APIRouteHandler } from "@/lib/services/api-route-handler";
+import { aiCacheOptimizationService } from "@/lib/services/ai-cache-optimization-service";
 
 export const GET = APIRouteHandler.createSimpleCachedGETHandler(
   async () => {
@@ -17,9 +17,8 @@ export const GET = APIRouteHandler.createSimpleCachedGETHandler(
       timestamp: new Date().toISOString(),
     });
 
-    // Get current cache optimization statistics
-    const optimizationMetrics = await getOptimizationMetrics();
-    const cacheHitRate = await getOverallCacheHitRate();
+    const optimizationMetrics = await aiCacheOptimizationService.getOptimizationMetrics();
+    const cacheHitRate = await aiCacheOptimizationService.getOverallCacheHitRate();
 
     return {
       ...optimizationMetrics,
@@ -27,308 +26,8 @@ export const GET = APIRouteHandler.createSimpleCachedGETHandler(
     };
   },
   {
-    ttl: 60, // 1 minute cache
+    ttl: 60,
     tags: ["ai-cache-optimization"],
-    varyBy: [], // Same response for all users
+    varyBy: [],
   },
 );
-
-/**
- * Get comprehensive cache optimization metrics
- */
-async function getOptimizationMetrics() {
-  const metrics = {
-    // Cache performance metrics
-    cachePerformance: await getCachePerformanceMetrics(),
-
-    // Cost optimization analytics
-    costOptimization: await getCostOptimizationAnalytics(),
-
-    // Pattern-based optimization effectiveness
-    patternOptimization: await getPatternOptimizationMetrics(),
-
-    // Time-based optimization efficiency
-    timeOptimization: await getTimeBasedOptimizationMetrics(),
-
-    // Overall optimization impact
-    overallImpact: await calculateOverallImpact(),
-  };
-
-  return metrics;
-}
-
-/**
- * Get cache performance metrics
- */
-async function getCachePerformanceMetrics() {
-  try {
-    // Get Redis cache performance statistics
-    const redisStats = await redisManager.executeWithFallback(
-      async (client) => {
-        const info = await client.info("memory");
-        const keyspace = await client.info("keyspace");
-
-        return {
-          memoryUsage: info,
-          keyspaceInfo: keyspace,
-        };
-      },
-    );
-
-    // Parse Redis memory information
-    const memoryInfo = parseRedisMemoryInfo(redisStats.memoryUsage);
-
-    return {
-      totalCacheKeys: await getTotalCacheKeys(),
-      memoryUsage: memoryInfo,
-      averageTTL: await getAverageTTL(),
-      cacheEfficiency: await calculateCacheEfficiency(),
-    };
-  } catch (error) {
-    logger.warn("Failed to get cache performance metrics", {
-      error: error instanceof Error ? error.message : String(error),
-    });
-
-    return {
-      totalCacheKeys: 0,
-      memoryUsage: { usedHuman: "N/A", used: 0 },
-      averageTTL: 0,
-      cacheEfficiency: 0,
-    };
-  }
-}
-
-/**
- * Get cost optimization analytics
- */
-async function getCostOptimizationAnalytics() {
-  try {
-    // Simulate cost optimization metrics based on cache hit rates
-    const cacheHitRate = await getOverallCacheHitRate();
-    const estimatedApiCallsPerHour = 100; // Example baseline
-    const costPerApiCall = 0.01; // Example cost
-
-    const savedCalls = Math.floor(
-      estimatedApiCallsPerHour * (cacheHitRate / 100),
-    );
-    const estimatedSavingsPerHour = savedCalls * costPerApiCall;
-
-    return {
-      cacheHitRate: `${cacheHitRate}%`,
-      estimatedApiCallsSaved: savedCalls,
-      estimatedCostSavings: {
-        perHour: `$${estimatedSavingsPerHour.toFixed(2)}`,
-        perDay: `$${(estimatedSavingsPerHour * 24).toFixed(2)}`,
-        perMonth: `$${(estimatedSavingsPerHour * 24 * 30).toFixed(2)}`,
-      },
-      optimisationFactor: calculateOptimizationFactor(cacheHitRate),
-    };
-  } catch (error) {
-    logger.warn("Failed to get cost optimization analytics", {
-      error: error instanceof Error ? error.message : String(error),
-    });
-
-    return {
-      cacheHitRate: "N/A",
-      estimatedApiCallsSaved: 0,
-      estimatedCostSavings: {
-        perHour: "$0.00",
-        perDay: "$0.00",
-        perMonth: "$0.00",
-      },
-      optimisationFactor: 1.0,
-    };
-  }
-}
-
-/**
- * Get pattern-based optimization metrics
- */
-async function getPatternOptimizationMetrics() {
-  try {
-    // Get cache metrics for different AI patterns
-    const patterns = [
-      "fintech",
-      "healthcare",
-      "saas",
-      "ecommerce",
-      "dashboard",
-    ];
-    const patternMetrics: Record<string, any> = {};
-
-    for (const pattern of patterns) {
-      const patternKeys = await getCacheKeysByPattern(pattern);
-      const avgTTL = await getAverageTTLForPattern(pattern);
-
-      patternMetrics[pattern] = {
-        cacheEntries: patternKeys.length,
-        averageTTL: avgTTL,
-        optimizationMultiplier: calculatePatternMultiplier(pattern),
-      };
-    }
-
-    return patternMetrics;
-  } catch (error) {
-    logger.warn("Failed to get pattern optimization metrics", {
-      error: error instanceof Error ? error.message : String(error),
-    });
-
-    return {};
-  }
-}
-
-/**
- * Get time-based optimization metrics
- */
-async function getTimeBasedOptimizationMetrics() {
-  const currentHour = new Date().getHours();
-  const currentMultiplier = getTimeBasedMultiplier(currentHour);
-
-  return {
-    currentHour,
-    currentMultiplier,
-    optimizationApplied: currentMultiplier !== 1.0,
-    timeCategory: getTimeCategory(currentHour),
-    projectedSavings: calculateTimeBasedSavings(currentMultiplier),
-  };
-}
-
-/**
- * Calculate overall optimization impact
- */
-async function calculateOverallImpact() {
-  const cacheHitRate = await getOverallCacheHitRate();
-  const optimizationFactor = calculateOptimizationFactor(cacheHitRate);
-
-  return {
-    performanceImprovement: `${((optimizationFactor - 1) * 100).toFixed(1)}%`,
-    costReduction: `${((cacheHitRate / 100) * 0.8 * 100).toFixed(1)}%`,
-    roiMultiplier: optimizationFactor.toFixed(2),
-    recommendation: getOptimizationRecommendation(cacheHitRate),
-  };
-}
-
-// Helper functions
-
-async function getTotalCacheKeys(): Promise<number> {
-  try {
-    const result = await redisManager.executeWithFallback(async (client) => {
-      const keys = await client.keys("cache:*");
-      return keys.length;
-    });
-    return result;
-  } catch {
-    return 0;
-  }
-}
-
-async function getOverallCacheHitRate(): Promise<number> {
-  // Simulate cache hit rate - in real implementation, this would track actual hits/misses
-  return 75; // 75% cache hit rate
-}
-
-function parseRedisMemoryInfo(info: string) {
-  const lines = info.split("\r\n");
-  const memoryData: Record<string, any> = {};
-
-  for (const line of lines) {
-    if (line.includes("used_memory_human:")) {
-      memoryData.usedHuman = line.split(":")[1];
-    }
-    if (line.includes("used_memory:")) {
-      memoryData.used = parseInt(line.split(":")[1]);
-    }
-  }
-
-  return memoryData;
-}
-
-async function getAverageTTL(): Promise<number> {
-  // Simulate average TTL calculation
-  return 3600; // 1 hour average
-}
-
-async function calculateCacheEfficiency(): Promise<number> {
-  const hitRate = await getOverallCacheHitRate();
-  const avgTTL = await getAverageTTL();
-
-  // Efficiency based on hit rate and TTL
-  return (hitRate * (avgTTL / 3600)) / 100;
-}
-
-function calculateOptimizationFactor(cacheHitRate: number): number {
-  // Rough calculation: each 10% hit rate = 15% improvement
-  return 1 + (cacheHitRate / 100) * 0.15;
-}
-
-function getTimeBasedMultiplier(hour: number): number {
-  if (hour >= 22 || hour <= 6) {
-    return 1.4; // Off-peak
-  } else if (hour >= 14 && hour <= 18) {
-    return 0.8; // Peak hours
-  }
-  return 1.0; // Normal
-}
-
-function getTimeCategory(hour: number): string {
-  if (hour >= 22 || hour <= 6) {
-    return "off-peak";
-  } else if (hour >= 14 && hour <= 18) {
-    return "peak";
-  }
-  return "normal";
-}
-
-function calculateTimeBasedSavings(multiplier: number): string {
-  if (multiplier > 1.0) {
-    return `${((multiplier - 1.0) * 100).toFixed(0)}% longer cache duration`;
-  } else if (multiplier < 1.0) {
-    return `${((1.0 - multiplier) * 100).toFixed(0)}% shorter cache duration`;
-  }
-  return "no change";
-}
-
-async function getCacheKeysByPattern(pattern: string): Promise<string[]> {
-  try {
-    return await redisManager.executeWithFallback(async (client) => {
-      return await client.keys(`*${pattern}*`);
-    });
-  } catch {
-    return [];
-  }
-}
-
-async function getAverageTTLForPattern(pattern: string): Promise<number> {
-  // Simulate pattern-specific TTL calculation
-  const patternTTLMap: Record<string, number> = {
-    fintech: 10800, // 3 hours
-    healthcare: 7200, // 2 hours
-    saas: 3600, // 1 hour
-    ecommerce: 3600, // 1 hour
-    dashboard: 1800, // 30 minutes
-  };
-
-  return patternTTLMap[pattern] || 1800;
-}
-
-function calculatePatternMultiplier(pattern: string): number {
-  const highValuePatterns = ["fintech", "healthcare"];
-  const mediumValuePatterns = ["saas", "ecommerce"];
-
-  if (highValuePatterns.includes(pattern)) {
-    return 1.6;
-  } else if (mediumValuePatterns.includes(pattern)) {
-    return 1.3;
-  }
-  return 1.1;
-}
-
-function getOptimizationRecommendation(hitRate: number): string {
-  if (hitRate < 60) {
-    return "Consider increasing cache TTL and implementing more aggressive warming strategies";
-  } else if (hitRate > 85) {
-    return "Excellent cache performance - consider fine-tuning TTL balance between freshness and efficiency";
-  } else {
-    return "Good cache performance - monitor for optimization opportunities";
-  }
-}
