+++ docs/architecture/blueprint.md
@@ -529,7 +529,7 @@ Agens must strictly follow these principles when generating code:
 // ✅ Centralized Type Definitions
 import { ServiceTypes } from './service-types.ts';
 
-// ✅ Specialized Services (31 total)
+// ✅ Specialized Services (32 total)
 - MonitoringService
 - MonitoringDashboardService
 - GitHubService
@@ -538,6 +538,7 @@ import { ServiceTypes } from './service-types.ts';
 - BlueprintEngine
 - EnterpriseThemeService
 - CacheOrchestrator (546 lines, 70% reduction - COMPLETED January 8, 2026)
+- AICacheOptimizationService (268 lines, LAYER SEPARATION - COMPLETED January 10, 2026)
 - MetricsCalculatorService
 - ErrorMonitoringService
 - PredictivePerformanceAnalyzer
