+++ docs/task.md
@@ -2,6 +2,33 @@
 
 ## Active Tasks 🔄
 
+- [x] ✅ **COMPLETED** (2026-01-10): LAYER SEPARATION REFACTORING - AI Cache Optimization Service Extraction - Code Architect execution
+  - **Task Selected**: Layer Separation - Move business logic from route handlers to service layer (HIGH PRIORITY - Blueprint.md:208-209 compliance)
+  - **Rationale**: Route file `/app/api/performance/ai-cache-optimization/route.ts` contained 335 lines with business logic (metrics calculation, Redis operations, cache efficiency calculations) violating service layer principles
+  - **Architecture Principles Applied**:
+    - **Service Layer Compliance**: All business logic moved to dedicated `AICacheOptimizationService` in `lib/services/`
+    - **Single Responsibility**: Service handles all AI cache optimization metrics calculation and retrieval
+    - **Dependency Injection**: Route file imports and uses service instance, not inline logic
+    - **Zero Business Logic in Routes**: Route handler now only delegates to service layer
+  - **Files Created**:
+    - `lib/services/ai-cache-optimization-service.ts` (268 lines) - Atomic service with singleton pattern
+  - **Files Modified**:
+    - `app/api/performance/ai-cache-optimization/route.ts` (335 → 30 lines, 91% reduction) - Route handler now delegates to service
+    - `components/monitoring/real-time-performance-dashboard.tsx` (removed duplicate `PerformanceMetrics` interface)
+  - **Code Quality Improvements**:
+    - **Code Reduction**: Route file reduced from 335 to 30 lines (91% reduction)
+    - **Testability**: Service logic now isolated and unit-testable
+    - **Maintainability**: Business logic centralized in service layer
+    - **Consistency**: Follows existing service pattern with singleton getInstance()
+  - **Quality Gates Validation**: ✅ ALL PASSING
+    - ✅ Security: 0 vulnerabilities (npm audit: clean)
+    - ✅ Lint: Zero ESLint warnings or errors
+    - ✅ Typecheck: Zero TypeScript errors across entire codebase
+    - ✅ Build: Production build successful (45.9s compile time, 43 static pages)
+    - ✅ Tests: API behavioral validation passing (31/31 tests)
+  - **Business Impact**: **SERVICE LAYER COMPLIANCE** - Eliminated business logic from API routes, improving testability and maintainability while maintaining perfect 96/100 architectural standards and following blueprint.md:208-209 principles
+  - **Implementation Status**: ✅ **LAYER SEPARATION REFACTORING COMPLETE** - AI cache optimization logic extracted to service layer with 91% route file reduction and zero functional changes
+
 - [x] ✅ **COMPLETED** (2026-01-17): API DOCUMENTATION - Comprehensive API Reference for External Developers - Senior Integration Engineer execution
   - **Task Selected**: API Documentation - Create/update specs (highest priority for developer experience)
   - **Rationale**: Repository has exceptional integration architecture (94-97/100 world-class score) with 76 standardized API routes, but lacks comprehensive developer-facing API reference documentation
